refactor: move clone utilities
