+++ src/turns/factories/turnActionFactory.js
@@ -2,7 +2,7 @@
  * @module turnActionFactory
  */
 
-import { freeze } from '../../utils/objectUtils.js';
+import { freeze } from '../../utils/cloneUtils.js';
 import { ITurnActionFactory } from '../ports/ITurnActionFactory.js';
 
 /**
