+++ src/utils/index.js
@@ -10,3 +10,4 @@ export { fetchWithRetry } from './httpUtils.js';
 export * from './loggerUtils.js';
 export * from './logHelpers.js';
 export * from './objectUtils.js';
+export * from './cloneUtils.js';
