+++ tests/unit/utils/expandActionArray.test.js
@@ -2,7 +2,7 @@ import {
   expandActionArray,
   expandMacros,
 } from '../../../src/utils/macroUtils.js';
-import { freeze } from '../../../src/utils/objectUtils.js';
+import { freeze } from '../../../src/utils/cloneUtils.js';
 import { describe, it, expect, jest } from '@jest/globals';
 
 const createRegistry = (macros) => ({
